+++ src/components/editgrid/EditGrid.js
@@ -109,14 +109,14 @@ export default class EditGridComponent extends NestedComponent {
       return this.addRow(true);
     }
     let tableClass = 'editgrid-listgroup list-group ';
-    _.each(['striped', 'bordered', 'hover', 'condensed'], (prop) => {
+    ['striped', 'bordered', 'hover', 'condensed'].forEach((prop) => {
       if (this.component[prop]) {
         tableClass += `table-${prop} `;
       }
     });
     const tableElement = this.ce('ul', { class: tableClass }, [
       this.headerElement = this.createHeader(),
-      this.rowElements = _.map(this.editRows, this.createRow.bind(this)),
+      this.rowElements = this.editRows.map(this.createRow.bind(this)),
       this.footerElement = this.createFooter(),
     ]);
 
@@ -127,9 +127,7 @@ export default class EditGridComponent extends NestedComponent {
       this.element.appendChild(tableElement);
     }
     //add open class to the element if any edit grid row is open
-    const isAnyRowOpen = _.some(this.editRows, function(row) {
-      return row.isOpen;
-    });
+    const isAnyRowOpen = this.editRows.some((row) => row.isOpen);
     if (isAnyRowOpen) {
       this.addClass(this.element, `formio-component-${this.component.type}-row-open`);
     }
@@ -502,6 +500,8 @@ export default class EditGridComponent extends NestedComponent {
 
   setValue(value) {
     if (!value) {
+      this.editRows = this.defaultValue;
+      this.buildTable();
       return;
     }
     if (!Array.isArray(value)) {
@@ -548,5 +548,17 @@ export default class EditGridComponent extends NestedComponent {
    */
   getValue() {
     return this.dataValue;
+  }
+
+  clearOnHide(show) {
+    super.clearOnHide(show);
+    if (!this.component.clearOnHide) {
+      // If some components set to clearOnHide we need to clear them.
+      this.buildTable();
+    }
+  }
+
+  restoreComponentsContext() {
+    return;
   }
 }