+++ Changelog.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [UNRELEASED]
+### Fixed
+ - Children components context for Container, DataGrid and EditGrid component.
 ### Added
  - Content property to HTML and Content components Logic.